requirements.txt
asgiref==3.7.2
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.0
pillow==10.2.0
pymongo==3.12.1
pytz==2024.1
sqlparse==0.2.4
tzdata==2024.1
